+++ src/apps/chifra/internal/transactions/options.go
@@ -31,6 +31,7 @@ type TransactionsOptions struct {
 	Reconcile      string                   `json:"reconcile,omitempty"`      // Please use --account_for option instead
 	AccountFor     string                   `json:"accountFor,omitempty"`     // Reconcile the transaction as per the provided address
 	Cache          bool                     `json:"cache,omitempty"`          // Force the results of the query into the tx cache (and the trace cache if applicable)
+	Decache        bool                     `json:"decache,omitempty"`        // Removes a transactions and any traces in the transaction from the cache
 	Source         bool                     `json:"source,omitempty"`         // Find the source of the funds sent to the receiver
 	Globals        globals.GlobalOptions    `json:"globals,omitempty"`        // The global options
 	BadFlag        error                    `json:"badFlag,omitempty"`        // An error flag if needed
@@ -49,6 +50,7 @@ func (opts *TransactionsOptions) testLog() {
 	logger.TestLog(len(opts.Flow) > 0, "Flow: ", opts.Flow)
 	logger.TestLog(len(opts.AccountFor) > 0, "AccountFor: ", opts.AccountFor)
 	logger.TestLog(opts.Cache, "Cache: ", opts.Cache)
+	logger.TestLog(opts.Decache, "Decache: ", opts.Decache)
 	logger.TestLog(opts.Source, "Source: ", opts.Source)
 	opts.Globals.TestLog()
 }
@@ -120,6 +122,8 @@ func transactionsFinishParseApi(w http.ResponseWriter, r *http.Request) *Transac
 			opts.AccountFor = value[0]
 		case "cache":
 			opts.Cache = true
+		case "decache":
+			opts.Decache = true
 		case "source":
 			opts.Source = true
 		default: